requirements.txt
about-time==4.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alive-progress==3.1.5
aniso8601==9.0.1
astor==0.8.1
async-timeout==4.0.3
attrs==24.2.0
autograd==1.7.0
autograd-gamma==0.5.0
bidict==0.23.1
blinker==1.8.2
captum==0.7.0
certifi==2024.8.30
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dicom==0.9.9.post1
early-stopping-pytorch==1.0.10
einops==0.8.0
et_xmlfile==2.0.0
feather==0.1.2
filelock==3.15.4
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SocketIO==5.4.1
fonttools==4.53.1
formulaic==1.0.2
frozenlist==1.4.1
fsspec==2024.6.1
grapheme==0.6.0
graphlib_backport==1.1.0
h11==0.14.0
histolab==0.6.0
huggingface-hub==0.26.2
idna==2.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.4
interface-meta==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lifelines==0.27.8
lmdb==1.4.1
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.1
numpy==1.24.4
opencv-python==4.10.0.84
openpyxl==3.1.5
openslide-python==1.2.0
packaging==24.1
pandas==2.0.3
pillow==10.4.0
progressbar==2.5
protobuf==5.28.2
psutil==6.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-slugify==5.0.2
python-socketio==5.11.4
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.2
reportlab==4.1.0
requests==2.25.1
safetensors==0.4.3
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.8.1
seaborn==0.13.2
shapely==2.0.6
simple-websocket==1.1.0
six==1.16.0
swin_transformer_v2 @ git+https://github.com/ChristophReich1996/Swin-Transformer-V2@7b74308bd560a5efabc0f9610d038dcf8e61226e
sympy==1.13.2
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.9.8
torch==2.4.0+cu118
torch-geometric==2.6.0
torch-scatter==2.0.7
torchaudio==2.4.0+cu118
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
yarl==1.11.1
zipp==3.20.1
